updating torch tensor for mask

+++ Pytorch/model/dataset.py
@@ -74,7 +74,7 @@ class Mindboggle_101(Dataset):
             # print("AFTER: ", sample['image'].shape, " msk ", sample['mask'].shape)
         
         sample['image'] = torch.from_numpy(sample['image'].copy().astype(np.float32)).permute(3, 0, 1, 2)
-        sample['mask'] = torch.from_numpy(sample['mask'].copy().astype(np.uint8)).permute(3, 0, 1, 2)
+        sample['mask'] = torch.from_numpy(sample['mask'].copy()).permute(3, 0, 1, 2)
 
         return sample
